+++ src/main/java/com/example/library/management/controller/UserController.java
@@ -2,10 +2,11 @@ package com.example.library.management.controller;
 
 import com.example.library.management.dto.UserRequest;
 import com.example.library.management.dto.UserResponse;
+import com.example.library.management.dto.login.LoginResponse;
+import com.example.library.management.security.AccessEnforcer;
 import com.example.library.management.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -21,13 +22,17 @@ public class UserController
     private static final Logger log = LoggerFactory.getLogger(UserController.class);
 
     private final UserService userService;
+    private final AccessEnforcer accessEnforcer;
 
-    public UserController(UserService userService) {
+    public UserController(UserService userService, AccessEnforcer accessEnforcer) {
         this.userService = userService;
+        this.accessEnforcer = accessEnforcer;
     }
 
     @GetMapping("/users")
     public List<UserResponse> getAllUsers() {
+        accessEnforcer.requireAdmin();
+
         log.debug("Received get all users request");
 
         return userService.getAllUsers();
@@ -41,14 +46,16 @@ public class UserController
     }
 
     @PostMapping("/users/login")
-    public ResponseEntity<UserResponse> loginUser(@RequestBody UserRequest loginDto) {
+    public LoginResponse<UserResponse> loginUser(@RequestBody UserRequest loginDto) {
         log.debug("Received user login request for email={}", loginDto.email());
 
-        return userService.getUser(loginDto.email(), loginDto.password());
+        return userService.login(loginDto.email(), loginDto.password());
     }
 
     @DeleteMapping("/users/{userId}")
     public void deleteUser(@PathVariable Long userId) {
+        accessEnforcer.requireAdmin();
+
         log.debug("Received delete user request for userId={}", userId);
 
         userService.deleteUser(userId);
